api const

+++ app/service/user.service.ts
@@ -3,7 +3,7 @@ import {Http, HTTP_PROVIDERS} from 'angular2/http';
 import {Observable} from 'rxjs/Observable';
 import 'rxjs/Rx';
 
-import {MyMatrixApi} from '../constants/apiConstants';
+import MyMatrixApi from '../constants/apiConstants';
 import {User} from '../models/user/user';
 import {StorageService} from './storage.service';
 // import {AuthorizationService} from './authorization.service';
@@ -16,7 +16,9 @@ export class UserService {
 
   public getUser(userId: number): void {
     const user = JSON.parse(this._storageService.getItem('MyMatrixUser'));
-    if (!user) {
+    if (user) {
+      this.emitUser(user);
+    } else {
       this._api.get(`${MyMatrixApi}/users/${userId}`)
         .map(res => <User>res.json())
         .subscribe(
@@ -24,8 +26,6 @@ export class UserService {
           err => console.log(err),
           () => console.log('User retrieval is completed')
         );
-    } else {
-      this.emitUser(user);
     }
   }
 
@@ -35,7 +35,7 @@ export class UserService {
 
   public updateUser(userId: number, user: User): void {
     const userBody = JSON.stringify(user);
-    this._api.put(`${MyMatrixApi}/users/${userId}`, userBody)
+    this._api.put(`${MyMatrixApi}/users/${userId}`, userBody, {headers: this._authorizationService.createAuthorizationHeader()})
       .subscribe(
         err => console.log('error: ', err),
         () => console.log('User updated')
@@ -43,7 +43,7 @@ export class UserService {
   }
 
   public deleteUser(userId: number): void {
-    this._api.delete(`${MyMatrixApi}/users/${userId}`)
+    this._api.delete(`${MyMatrixApi}/users/${userId}`, {headers: this._authorizationService.createAuthorizationHeader()})
       .subscribe(
         err => console.log('error: ', err),
         () => console.log('User updated')
